% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickEffectSize.R
\name{QuickEffectSize}
\alias{QuickEffectSize}
\alias{qes}
\title{QuickEffectSize (qes)}
\usage{
QuickEffectSize(zelig.model, iv.var, sim.n = 100, range.n = 100,
  custom.range, return.pdata = FALSE, xlab = "IV", ylab = "DV",
  coord.ylim, set.covar = NULL, ...)
}
\arguments{
\item{zelig.model}{Fitted zelig model}

\item{iv.var}{Independent variable.}

\item{sim.n}{Number of simulations for each value of iv.var simulated. Defaults to 100.}

\item{range.n}{Number of different values of iv.var to simulate. Defaults to 100.}

\item{custom.range}{Vector of two values specifying a range within which different values of iv.var should be simulated. (Optional)}

\item{return.pdata}{Should the data to construct the plot be returned instead of the plot itself? Defaults to FALSE.}

\item{coord.ylim}{Sets limits on what part of the x.axis to display through ggplots coord_cartesian() function. (Optional)}

\item{set.covar}{Option to specify values of other predictors in simulations. (Optional)}
}
\description{
This package visualizes effect sizes. It relies on the Zelig package.
}
\examples{
dat <- data.frame(y = rnorm(100), x1 = rnorm(100), x2 = rnorm(100))
dat$x3 <- dat$y + rnorm(100)
library(Zelig)
qes(zelig(y ~ x1 + x2 + x3, data = dat, model = "normal"), iv.var = "x3", xlab = "Using qes", ylab = "Productivity")
}
\keyword{QuickEffectSize}
\keyword{Zelig}
\keyword{effect}
\keyword{plot}
\keyword{qes}
